// src/agent/model/providers/copilot/responses/openai-responses-api-types.ts
export type OpenAIResponsesFileSearchToolComparisonFilter = {
  key: string
  type: 'eq' | 'ne' | 'gt' | 'gte' | 'lt' | 'lte'
  value: string | number | boolean
}

export type OpenAIResponsesFileSearchToolCompoundFilter = {
  type: 'and' | 'or'
  filters: Array<OpenAIResponsesFileSearchToolComparisonFilter | OpenAIResponsesFileSearchToolCompoundFilter>
}